refactor: split models into loaders and updaters
