+++ src/palletjack/__init__.py
@@ -1,4 +1,5 @@
 """A library for updating AGOL feature services with data from SFTP sources
 """
 
-from .models import ColorRampReclassifier, FeatureServiceInlineUpdater, FeatureServiceOverwriter, SFTPLoader
+from .loaders import GSheetLoader, SFTPLoader
+from .updaters import ColorRampReclassifier, FeatureServiceInlineUpdater, FeatureServiceOverwriter
